Fix coredumper in case of RELRO executable.
Thanks to the issue description and the proposed patch in
https://code.google.com/p/google-coredumper/issues/detail?id=11
it was easy to find the cause and fix it. However the patch from
the official coredumper tracker isn't so good and would dump all
the content of all linked libraries in the core dump, which would
increase the core size of hundreds of MB in the case of bi
applications. This fix targets just the right mapping and dump
it.